Move buildURL and related methods to a mixin
This is my first step towards emberjs/rfcs#4.

The intent here is to separate url building from the RESTAdapter so that
it can eventually be included in other adapters and hold the logic for the
path templating described in emberjs/rfcs#4.
